+++ boardctl/plugins/executors/go.py
@@ -3,5 +3,5 @@
 NAME = 'go'
 
 
-def build_cmd(addr):
+def build_cmd(addr, t=None):
     return f'go {addr}'