+++ boardctl/plugins/executors/source.py
@@ -3,5 +3,5 @@
 NAME = 'source'
 
 
-def build_cmd(addr):
+def build_cmd(addr, t=None):
     return f'source {addr}'